added resolvers for default branchs and collaborators
